requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-Limiter==3.7.0
APScheduler==3.10.4
python-dotenv==1.0.1
marshmallow==3.21.3
marshmallow-sqlalchemy==0.29.0
passlib[bcrypt]==1.7.4
google-generativeai==0.7.2
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.13
moviepy==1.0.3
requests==2.32.3
itsdangerous==2.2.0